+++ src/activitypub/actors/helpers.rs
@@ -23,7 +23,7 @@ use crate::models::{
     },
 };
 
-const ACTOR_IMAGE_MAX_SIZE: usize = 5 * 1000 * 1000; // 5 MB
+pub const ACTOR_IMAGE_MAX_SIZE: usize = 5 * 1000 * 1000; // 5 MB
 
 async fn fetch_actor_images(
     instance: &Instance,