+++ src/mastodon_api/media/views.rs
@@ -26,6 +26,7 @@ async fn create_attachment_view(
         &attachment_data.file,
         attachment_data.media_type.clone(),
         &config.media_dir(),
+        config.limits.media.file_size_limit,
         None,
     )?;
     let db_attachment = create_attachment(