+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Save latest ethereum block number to database instead of file.
+- Removed hardcoded upload size limit.
 
 ### Deprecated